Extract all mercury payload to one file
