--- mercury/model/build_sudt_issue_transactions.go
@@ -1,14 +1,0 @@
-package model
-
-import (
-	"github.com/nervosnetwork/ckb-sdk-go/mercury/model/req"
-)
-
-type BuildSudtIssueTransactionPayload struct {
-	Owner                  string                 `json:"owner"`
-	From                   []*req.Item            `json:"from"`
-	To                     []*ToInfo              `json:"to"`
-	OutputCapacityProvider OutputCapacityProvider `json:"output_capacity_provider,omitempty"`
-	FeeRate                uint64                 `json:"fee_rate,omitempty"`
-	Since                  *SinceConfig           `json:"since,omitempty"`
-}
