--- mercury/model/dao_withdraw_payload.go
@@ -1,8 +1,0 @@
-package model
-
-import "github.com/nervosnetwork/ckb-sdk-go/mercury/model/req"
-
-type DaoWithdrawPayload struct {
-	From    []*req.Item `json:"from"`
-	FeeRate uint64      `json:"fee_rate,omitempty"`
-}
