--- mercury/model/get_spent_transaction_payload.go
@@ -1,17 +1,0 @@
-package model
-
-import (
-	"github.com/nervosnetwork/ckb-sdk-go/types"
-)
-
-type GetSpentTransactionPayload struct {
-	OutPoint      types.OutPoint `json:"outpoint"`
-	StructureType StructureType  `json:"structure_type"`
-}
-
-type StructureType string
-
-const (
-	Native      StructureType = "Native"
-	DoubleEntry StructureType = "DoubleEntry"
-)
